requirements.txt
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
coloredlogs==15.0.1
flatbuffers==25.9.23
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lazy_loader==0.4
llvmlite==0.43.0
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
onnxruntime==1.19.2
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
platformdirs==4.4.0
pooch==1.8.2
protobuf==6.33.0
PyMatting==1.1.14
referencing==0.36.2
rembg==2.0.61
requests==2.32.5
rpds-py==0.27.1
scikit-image==0.24.0
scipy==1.13.1
sympy==1.14.0
tifffile==2024.8.30
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
